pyproject.toml
[tool.poetry]
name = "killthenoise-backend"
version = "0.1.0"
description = "Backend API for KillTheNoise, built with FastAPI"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
asyncpg = "^0.29.0"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.29"}
alembic = "^1.13.0"
supabase = "^2.2.0"
pydantic-extra-types = "^2.6.0"
multiprocessing-logging = "^0.2"
psycopg2-binary = "^2.9"
anthropic = "^0.18.1"

[tool.poetry.group.dev.dependencies]
black = "^24.3"
isort = "^5.13"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pytest = "^8.1"
pre-commit = "^3.7"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api" 